+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2019-1-1
+### Added
+- Added the ability for the bottom panel to switch between notification and
+chat.
+- Added the ability for the bottom panel to be minimised and expanded.
+- Added CSS animation to the bottom panel when it is expanded / minimised.
+
+### Fixed
+- Fixed a bug with the local storage where the notifications were stored as
+an entire string instead of individual sentences.
+
 ## [0.3.0] - 2019-1-1
 ### Added
 - Created a bottom panel to house the notification and chat functionality that